Remove storage dependency on iris-grid
