+++ packages/iris-grid/src/filters/FilterOperator.ts
@@ -1,12 +1,2 @@
-export class FilterOperator {
-  static readonly not = 'not';
-
-  static readonly and = 'and';
-
-  static readonly or = 'or';
-}
-
-export type FilterOperatorValue = typeof FilterOperator[Exclude<
-  keyof typeof FilterOperator,
-  'prototype'
->];
+export { FilterOperator } from '@deephaven/storage';
+export type { FilterOperatorValue } from '@deephaven/storage';
