+++ packages/storage/src/filters/index.ts
@@ -1,0 +1,2 @@
+export * from './FilterOperator';
+export * from './FilterType';
