+++ packages/storage/src/index.ts
@@ -1,3 +1,4 @@
+export * from './filters';
 export * from './Storage';
 export * from './StorageTableViewportUpdater';
 export * from './StorageUtils';
