+++ packages/storage/tsconfig.json
@@ -8,9 +8,6 @@
   "exclude": ["node_modules", "src/**/*.test.*", "src/**/__mocks__/*"],
   "references": [
     {
-      "path": "../iris-grid"
-    },
-    {
       "path": "../log"
     }
   ]
